Move all case diff logic into casediff module
